LU-7091 mdd: refresh nlink after update linkea
It should refresh nlink after update linkea, because
it might decrease the nlink of source object,
otherwise it might try to delete the stripeEA with
wrong nlink.

And also before migration, it should check if all
of the parents of the migrating object are in the
target MDT, only checking if these parents parent
are remote is not enough.

Signed-off-by: wang di <di.wang@intel.com>
Change-Id: Icfb4ffc666855f0c7f35f004ecc864c422610135