LU-6215 ofed: strengthen ib_create_cq in 4.2+ kernels
External infinband stacks tend to keep their kernel
headers in sync with the latest kernels but this is
not the case with the driver code itself. Since this
is the case we can not trust testing for API changes
using OFED headers but need to be test driver code
API changes. Update the autoconf test to see if
ib_create_cq() uses struct ib_cq_init_attr.

Change-Id: I9875b6787723589b9cbdb55ef088b146b6900b2a
Signed-off-by: James Simmons <uja.ornl@yahoo.com>