LU-7042 ofed: Handle OFED 3.18 packaging definitions
Starting with OFED 3.18 the OFED package started to use
AC_PACKAGE_NAME macro which is also used by Lustre. This
lead to clash between the two in the ko2iblnd driver.
The solution is to undefine the PACKAGE_XXX values set
in compact-2.6.h so the proper lustre PACKAGE_XXX values
are used.

Change-Id: I64ea2709600165a01c2f1d086fceeca6778ae53f
Signed-off-by: James Simmons <uja.ornl@yahoo.com>