LU-5614 build: use %kernel_module_package in rpm spec
Replace module packages definitions with a call to the vendor defined
%kernel_module_package macro.
kmod package has ability to check and copy modules for the new added
kernel. The package includes dependencies on a kernel symbols version.
In most cases this is enough for a minor security kernel updates.

This patch changes the Lustre modules location to
{extra|updates}/lustre/{fs|net}/. Kmod utils/scripts
requires this modification, cause they use %name for postinstall
scripts.

Change-Id: I0aa3ba3ee87dde894c3432ee8e6b05cedcbb8bc1
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>