LU-5614 build: use %kernel_module_package in rpm spec
Replace module packages definitions with a call to the vendor defined
%kernel_module_package macro.
kmod package has ability to check and copy modules for the new added
kernel. The package includes dependencies on a kernel symbols version.
In most cases this is enough for a minor security kernel updates.

This patch changes the Lustre modules location to
{extra|updates}/lustre/{fs|net}/. Kmod utils/scripts
requires this modification, cause they use %name for postinstall
scripts.

Signed-off-by: Stephen Champion <schamp@sgi.com>
Signed-off-by: Alexander Boyko <alexander.boyko@seagate.com>
Change-Id: Ic5351f4bda94eae80eaf349c8490a2c758b15c80
Signed-off-by: Minh Diep <minh.diep@intel.com>
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>