LU-5614 build: use %kernel_module_package in rpm spec
This patch is for Morrone's build testing only.  Not meant to land.

Replace module packages definitions with a call to the vendor defined
%kernel_module_package macro.

The files that contain the file listing for each %kernel_module_package
call are called out with their own Source<N>: lines in the spec file
to make them available to the spec file in advance of %prep unpacking
the tarball.

Change-Id: I7f22bd7c720c4d5d6079b3821e757d73c833d95e
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Test-Parameters: fortestonly