LU-5125 lu: weaken assertion in lu_object_attr()
In lu_object_attr() assert that the object exists or is remote rather
than just exists. In the OSP layer the exists flag may be cleared by
certain operations and there is no locking to make it safe for other
threads to use lu_object_attr() even if they first check that the
object exists.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I5777d16e4e72ac2080a01e38adca2eaba03fd8b7