LU-0000 osc: identify pages pinned for bulk
Adding cl_page::cp_pinned to track pages if they are pinned
for bulk transfer. Also, if a page is being held for bulk transfer,
it shouldn't be reclaimed by OSC LRU.

A policy is added to handle those pinned pages. In ll_releasepage(),
if the page is currently pinned by bulk, a sync request will be sent
to OST to commit the transaction. Syncing a single object shouldn't
be a problem for ldiskfs and it will be good for ZFS also after zil
is in place.

unstable_stats under LLITE is removed.

unstable check still exists under OSC only statistic purpose.

Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: I4f4d1ca9520bb44a15a4dc7fef568846a9ab944b