LU-7311 osd: ext4_bread has one less parameter in kernel 3.18
The returned error code was the 5th parameter to ext4_bread until the
kernel 3.18. It was then removed an integrated into the returned
pointer. Create a shim to emulate the newer version of ext4_bread with
older kernels.

To use the shim, the following must be added to ext4.h in one of the
ldiskfs patches, likely ext4-misc.patch.

  #define HAVE_EXT4_BREAD_4ARGS 1

Signed-off-by: frank zago <fzago@cray.com>
Change-Id: I3dacd8b24ae7a3a823f5cc662b2ee3627f8fcda7