LU-7543 obd: reserve some connection flags
The following four connections flags are used for some purposes
in related patches that will be landed to master in the future.
Land the connection flags to master earlier for reserving slots
to avoid potential conflict with others.

  OBD_CONNECT_SUBTREE: LU-28 mounting of filesystem from MDS
  http://review.whamcloud.com/5007

  OBD_CONNECT_OBDOPACK: LU-4215 optimize OUT protocol
  http://review.whamcloud.com15336

  OBD_CONNECT_LOCK_AHEAD: LU-6917 LDLM lock ahead
  http://review.whamcloud.com/13564

  OBD_CONNECT_FLAGS2:
  LU-7543 reserve obd_connect_flags2 for future flags

The last flag is a feature for the client and server to use
obd_connect_flags2 to communicate future feature flags. The
client should set this flag whenever any flags in that field
are requested, and the server should mask unsupported features
from this field (assuming it understands OBD_CONNECT_FLAGS2).
When checking if an OBD_CONNECT2_xxxx feature is supported,
the client/server needs to firstly check if OBD_CONNECT_FLAGS2
is supported, since this field is also beyond the end of the
old obd_connect_data.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ie0ad6970c5726df025693f48a2d0cce270b81360