LU-7039 llog: update llog header and size
Once update request fails due to eviction or other failures,
all of update request in the sending list should return fail,
because after the failure, the update log in the following
request will have wrong llog bitmap. So once this happens,it
will

1. invalidate all of requests in the sending list, and if the
failure happens in replay, then all of replay and resend
requests needs to be invalidated as well, otherwise the llog
will also be corrupted.
2. Update the llog header from remote target.
3. Then Sending list can accept new request.

Also a few other fixes for llog corruption

1. Keep llog size updated in OSP cache for every llog write,
otherwise the size might be inconsistency.

2. rollback the lgh_index and bitmap once add new records
fails.

Add test_117 and update test_71a to verify the failure.

Signed-off-by: Di Wang <di.wang@intel.com>
Change-Id: I2d3a700d3363867ac60aeb6b7641eceb65dfe12a