LU-5216 hsm: restore doesn't restart for new copytool
This patch handles:

1. Unexpected client (data mover node) eviction could
cause on going hsm requests to be stuck in "STARTED"
state as the copy tool running on the data mover node
is no more available. This patch unregisters the
copy tool and cancels all the requests on the copy
tool agent.

2. The way to stop copy tool is to KILL it. In this
case also, as above explained, on going hsm operations
get stuck in "STARTED" state and if any other copy tool
tries to do any hsm activity  on that file will not be
processed as the hsm operation on that file still going
on. Such stuck operations will be cleared after request
time out but we need handle this immediately once the
copy tool agent dies, so that other copy tool can process
smoothly. Patch will handle this situation by waiting for
the on going hsm activity to complete and free the request.
We need to wait for the request complete and return to
make sure that the hsm operation should not be blocked.

Adding test_61 to sanity-hsm.sh to verify the fix

Seagate-bug-id: MRP-2464
Signed-off-by: vinayakswami hariharmath <vinayakswami.hariharmath@seagate.com>
Change-Id: I611775f93029805644dbf486050e59c7836c91e1