LU-7053 osd: don't lookup object at insert
the idea is to cache FID->ino/type mapping in per-thread cache
at declaration/object creation. then insert can find that information
and don't lookup object in LU/OI. this should avoid potential deadlock
with lu_object_find() and iget(). also, this should improve performance
as in the majority of cases required data is filled locally by create.

the patch is still in a proto stage and to be proved by maloo.

todo:
 - dynamic cache size to support wide striping (many creates a time)
 - unification with existing OI cache?
 - few places where error handling is missing

Change-Id: Ifa66c2d074f04e47d0d85b735f57dc506aa65f4c
Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>