LU-3323 lnet: Add support to restrict connections by NID
This patch adds support for an additional lnet module parameter called
permitted_nids which takes a list of NID ranges identical to root
squash.  The variable serves as a white-list of NID ranges which are
permitted to connect (active or passively) to a host.

An interface has been created to set/get the values for permitted_nids
through ioctls using the lctl commmand.

Upon updating the the permitted_nids list the peer hash list will be
traversed, dropping any peers which do not meet the new criteria.
This is implemented through an LND function pointer called lnet_drop
which must be supported by the LND to remove peers.  It has been added
for the ko2iblnd and ksocklnd LNDs.

In addition a new macro CDEBUG_AUDIT has been added to utilize the
kernel auditing facility and generate messages for these security
relevant changes as well as active and passive connections created
by the ksocklnd and ko2iblnd LNDs.  This auditing is only enabled
when Lustre is compiled with the --enable-auditing option to
configure.

Signed-off-by: Jeremy Filizetti <jeremy.filizetti@gmail.com>
Change-Id: I34dce9db66fe631a57fae9ee3510aa25dbb961c8