LU-3538 cos: reuse cos lock if compatible
Currently MDT local lock can't be reused, which may cause a huge
number of COS locks for the same object created. To avoid this,
when MDT reint operation locks an object, it will first try to
match a compatible COS lock, if found, this lock is used,
otherwise enqueue a new lock.

This patch contains following changes:
* when MDT reint operation locks an object, it will first try to
  match a compatible COS lock, if found, convert it to LCK_EX
  so that concurrent lock to this object can be detected, else
  enqueue a new local lock like before.
* for above reused COS lock, if MDT reint operation failed, it
  needs to be downgraded to LCK_COS, otherwise commit-on-sharing
  won't work.
* limit PDO hash size to 255, so that there are at most 255 COS
  hash locks for a directory, this can improve directory COS lock
  reuse.

Benchmark result with createmany/unlinkmany is as follows:
        mkdir rmdir open unlink mknod unlink (ops/sec)
2.6     1194 1310 1314 1185 2242 1396
master   978 1166  937 1028 1681 1202
cos      930 1161  918 1018 1691 1202
reuse    948 1162  925 1013 1676 1213

10 createmany/unlinkmany processes running on local client
(on MDS), 4M dirs/files created/unlinked, and the numbers are
average of 10 processes.
* for 2.6, each process is running on a separate mountpoint.
* cos is for commit I91928d097cbb26bd1e1089c3f8851ac6a6440a69.
* reuse is current commit.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I4a85fb2aa5035210bf12c5ba8dc49ba0e0e61c84