LU-3538 cos: reuse cos lock if compatible
Currently MDT local lock can't be reused, which may cause a huge
number of COS locks for the same object created. To avoid this,
when MDT reint operation locks an object, it will first try to
match a compatible COS lock, if found, this lock is used,
otherwise enqueue a new lock.

This patch contains following changes:
* when MDT reint operation locks an object, it will first try to
  match a compatible COS lock, if found, convert it to LCK_EX
  so that concurrent lock to this object can be detected, else
  enqueue a new local lock like before.
* for above reused COS lock, if MDT reint operation failed, it
  needs to be downgraded to LCK_COS, otherwise commit-on-sharing
  won't work.
* limit PDO hash size to 255, so that there are at most 255 COS
  hash locks for a directory, this can improve directory COS lock
  reuse.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I4a85fb2aa5035210bf12c5ba8dc49ba0e0e61c84