LU-7518 build: Remove the Phi accelerator-specific packaging
Intel needed to build Lustre for its non-self-hosted accelerator version
of the Phi processor. It turns out that the OS for that is packaged in some
weird non-standard way. Intel added support into Lustre to rpm package it.
While those additions use words like "cross_build", making it sound like it
might be generic, it really isn't.

All of that is really messy and non-standard, and it is making far too
difficult to package Lustre for normal rpm-based distributions.

Hopefully this won't be too much of a problem for Intel, because Phi
processors are now self-hosting.

This is in large part a revert of commit
  426194de03ad47497e99d281d86b5cdfc5142148:
"LU-2784 build: Provide RPMs build for Xeon Phi(TM) card"

Change-Id: Ie5d655f1da3e17eae6edebb5f4d6fd6e7d58811f
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>