LU-5282 mdc: fix panic at mdc_free_open()
Assertion was happened for open request when rq_replay is set
to 1.
    ASSERTION(mod->mod_open_req->rq_replay == 0)
But this situation is not fatal for client, and could happened
when mdc_close() failed.
The fix allow to free such requests. If mdc_close fail, MDS doesn`t
receive close request from client. And in a worst case client would
be evicted.

Signed-off-by: Alexander Boyko <alexander.boyko@seagate.com>
Change-Id: I5f98901f633355849fc107149eadc9cf171819af