LU-6654 osd: add RCU lock on osd_thread_info
The osd_iobuf in osd_thread_info is used by a thread
other than the one which owns the osd_thread_info.

The existing concurrency control (checking dr_numreqs) has
a race condition where dr_numreqs is decremented, which
opens the path for the other thread to eventually free
osd_thread_info, and then the first thread calls wakeup
on a wait queue which is inside the now freed osd_thread_info.

This patch adds RCU lock in "osd_thread_info" to protect
the osd_thread_info to NOT be freed when the other thread
is still using it.

Change-Id: I58a7181d3ee3b99e22e79a060173d49d7bcd71b5
Signed-off-by: Hongchao Zhang <hongchao.zhang@intel.com>