LU-7528 test: Deregister changelog client on test fail.
The tests which are related to changelog did not deregister
changelog when any failure was encountered, because of which
any changes afterward were also recorded leading to huge
test stdouts. I have added the functions error_and_deregister
which will first deregister thechangelog client created by
the test before exiting. This function is to be used in place
of error in any test that registers a changelog client.

Seagate-bug-id: MRP-3063
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>
Change-Id: I7f26f266ba8bda294b75ff5619d95c26704fd83f