LU-3953 build: Only chmod dkms.mkconf once
With AC_CONFIG_FILES, the "commands" parameter (the second one) is
applied for _each_ command, not just once.  That means the existing
chmod command was run many times, and several of the times it runs
it complains because the dkms.mkconf file does not yet exist.

This patch fixes that by giving the dkms.mkconf file its own
AC_CONFIG_FILES macro.

Change-Id: Ic71cc5d8c3555d28ff16efa23d564dce28662443
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>