LU-7734 lnet: Multi-Rail local NI split
This patch allows the configuration of multiple NIs under one Net.
It is now possible to have multiple NIDs on the same network:
Ex: <ip1>@tcp, <ip2>@tcp.
This can be configured using the following syntax:
Ex: tcp(eth0, eth1)

The data structures for the example above can be visualized
as follows

               NET(tcp)
                |
        -----------------
        |               |
      NI(eth0)        NI(eth1)

For more details refer to the Mult-Rail Requirements and HLD
documents

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Id7c73b9b811a3082b61e53b9e9f95743188cbd51