LU-7356 build: handle static/shared configure option for mount plugin
Currently if you build lustre with configure options
--disable-shared and/or --disable-static it will fail
to build in the case of the mount plugins. The mount
plugins are always built to be sharable code using
libtool but this is not compatiable with the confiure
--disable-share option. The work around is to not use
libtool but instead build the plugins as non installed
static libraries and using build tools turn them into
dynamic code.

Change-Id: I0eec25c8bce71c0e5c533ff3dfbe752677f37a26
Signed-off-by: James Simmons <uja.ornl@yahoo.com>