LU-2675 utils: restore lshowmount utility
The lshowmount utility was removed in commit b5a7260ae8f as being
obsolete, but it was not as unused as previously thought.

Restore lshowmount.c, lshowmount.8, nidlist.c, and nidlist.h  from
history, with minor updates to the Makefiles and .gitignore to avoid
conflicts in context.

No testing has been done on this utility, so it may not even build
at this point.  This is just a starting point for reviving the tool.
The Test-Parameters line and this paragraph can be removed once
someone has actually done some testing on this utility, and written
a test for conf-sanity that verifies it continues to work.

Test-Parameters: fortestonly
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Ibf6f72684d4dcaa95e0366b4fde74386893ebbe5