LU-7206 osp: Fix for LASSERT on otr_wakeup_list.
From the LASSERT it is clear that the variables otr_refcount
which represents no. of opd_devices pointing to it and
otr_wakeup_list which is the list of devices with active
transactions are out of sync. The vmcore shows that the
refcount is 0 but there are still active devices in the
wakeup list.

It was observed that all the members of the tracker including
the refcount and otr_wakeup_list would need to be protected
using the appropriate locks at all places which was not done
currently. This will protect the variable better avoiding
them to go out of sync as it is happening now.

Change-Id: Ia2d97775b5066e75f76d8437e0c8fe04f7f47530
Seagate-bug-id: MRP-2970
Signed-off-by: Kirtankumar Krishna Shetty <kirtan.shetty@seagate.com>