LU-7579 osd: move ORPHAN/DEAD flag to OSD
When a directory is unlinked, and becomes an orphan,
let's move DEAD/ORPHAN flag to OSD, so the remote
directory can retrieve these flags by OSP attr_get.

In the mean time, it will remove LUSTRE_SLAVE_DEAD_FL
flag for dead stripes of striped directory, so all of
dead directories (stripes) will marked with ORPHAN
flag in LMA.

Also in osp_xattr, it should retrieve the return
value from the reply, which is the real length of
EA, instead of using reply buffer length.

Signed-off-by: Di Wang <di.wang@intel.com>
Change-Id: I32f864cdcc72b460b4c2169d8e34211522b9c266