LU-7579 mdd: Add ORPHAN/DEAD flag in LMA
When a directory is unlinked, and becomes an orphan,
let's store the DEAD/ORPHAN flag in LMA, so the remote
directory can retrieve these flags by OSP xattr_get.

Also in osp_xattr, it should retrieve the return
value from the reply, which is the real length of
EA, instead of using reply buffer length.

Signed-off-by: Di Wang <di.wang@intel.com>
Change-Id: I32f864cdcc72b460b4c2169d8e34211522b9c266