LU-7613 dcache: changes made to ll_splice_inode to avoid dcache corruption.
ll_splice_alias is similar to d_splice_alias which
splice a disconnected dentry into the tree if one
exists.If inode is a directory and has a 'disconnected'
dentry (i.e.DCACHE_DISCONNECTED) then move that in
place of the given dentry  and return it.
This is needed in the lookup routine of any filesystem
that is exportable  (via knfsd) so that we can build
dcache paths to directories effectively.As this function
is valid only for building dcache path for directories
and should not be used for non-directories which might
result in corrupted dcache. So before building dcache path
for inode a simple check is added to confirm if the inode
is directory.

Seagate-bug-id: MRP-2739
Change-Id: Ieb9dabab0784bdb3c52e2cb32be1a766ffe55313
Signed-off-by: Lokesh Nagappa Jaliminche <lokesh.jaliminche@seagate.com>