LU-5030 tests: handle missing common_name for upstream client
The upstream stream client has moved from procfs to sysfs
except for llite.*.lov.common_name which does not exist
for the upstream client. The solution is to use lfs getname
which actually grabs the common_name using an ioctl and then
removes the clilov from the name before returning the result
to userland. For our use case we just add clilov back into
the name. Some of the scripts use the common_name only to
retrieve the number of OSTs. This can be done using lfs osts
instead.

Change-Id: I7fbb7348e009cb9f77ba9847aeb4879dc302e4ee
Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>