LU-5030 utils: migrate most lustre utils to use cfs_get_paths()
Move all the lustre utilities and all the test
applications except for ltrack_stat to use
cfs_get_paths() instead of accessing the
proc file system directly. This automatically
enables support for sysfs used in the upstream
client. The llite stats exist in two places so
a more complex solution is needed for ltrack_stat.

Change-Id: Iecb4a269c00ac8a5e0633d69d0370e9881ee6e33
Signed-off-by: Dmitry Eremin <dmitry.eremin@intel.com>
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Signed-off-by: Wang Chao <chao.ornl@gmail.com>