LU-7656 tests: tar a temporary folder
tar on /etc folder leads to various "No such
file or directory" failures for different files
in the /etc folder.

Since the main purpose of this test is to check
mdt recovery while tar runs in a loop, instead
of performing a tar on the /etc folder, the fix
is to create a temporary folder with n files in
it and perform a tar on this temporary folder.
After the tar process is killed, remove this
temporary folder.

Seagate-bug-id: MRP-2925
Signed-off-by: Noopur Maheshwari <noopur.maheshwari@seagate.com>
Change-Id: I62c4f9855c028094675bde222b630e6c847b95ac