LU-7734 lnet: configure peers from DLC
This patch adds the ability to configure peers from the DLC
interface.

When a peer is added a primary NID must be provided. If none is
provided then the first NID in the list of NIDs shall be used
as the primary NID.

Basic error checking is done at the DLC level, to ensure properly
formatted NIDs, however, if a NID is a duplicate, this will be
detected when adding it in the kernel. Operation is halted, which
means some peer NIDs might have already been added, but not the
entire set. It's the role of the caller to backtrack and remove that
peer that failed to add.

When deleting a peer the primary NID must be provided. If the primary
NID is not provided the operation fails and no changes occur in the
system. If only the primary NID is provided the entire peer is
removed.  If a set of NIDs are provided each one will be removed
from the peer identified by the peer NID in turn.

The existing show peer credits API can be used to show peer
information.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Iaf588a062b44d74305aa9aa7d31c7341c6c384b9